# backend/docs/postgres-reference/openapi.design.yaml
openapi: 3.0.3
info:
  title: School Admission Management API
  version: 1.0.0
  description: |
    REST API for enquiries, applications, workflow transitions, assessments,
    payments and reporting.

    **Conventions**
    - Base path `/api/v1`. JSON only. Timestamps are ISO-8601 UTC.
    - Auth: OAuth2 / OIDC bearer JWT. Guardians sign in with OTP; the JWT carries a `roles` claim.
    - Errors use RFC 9457 `application/problem+json`.
    - Mutating calls accept an `Idempotency-Key` header; payment calls require it.
    - Applications use optimistic locking: send `If-Match: <version>` on writes; a stale value returns `409`.
    - Lists are cursor-paginated: `?limit=25&cursor=...`.
servers:
  - url: https://api.example-school.in/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /applications:
    get:
      tags: [Applications]
      summary: Search applications (staff) or list my applications (guardian)
      parameters:
        - { name: stage, in: query, schema: { type: string, example: REGISTRATION } }
        - { name: classCode, in: query, schema: { type: string } }
        - { name: academicYear, in: query, schema: { type: string, example: 2027-28 } }
        - { name: assignedTo, in: query, schema: { type: string, format: uuid } }
        - { name: q, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Page of applications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Application' }
    post:
      tags: [Applications]
      summary: Start a draft application (guardian)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [classCode, academicYear]
              properties:
                classCode: { type: string }
                academicYear: { type: string }
                campusId: { type: string, format: uuid }
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
  /applications/{applicationId}:
    parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
    get:
      tags: [Applications]
      summary: Get an application with its checklist and next actions
      responses:
        '200':
          description: The application
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationDetail' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Applications]
      summary: Save draft answers (partial update)
      parameters:
        - { name: If-Match, in: header, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicant: { $ref: '#/components/schemas/Applicant' }
                guardians:
                  type: array
                  items: { $ref: '#/components/schemas/Guardian' }
                responses:
                  type: object
                  additionalProperties: true
                  description: Answers keyed by form field key
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /applications/{applicationId}/submit:
    post:
      tags: [Applications]
      summary: Submit the application form (runs the ENQUIRY to APPLICATION transition)
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
      responses:
        '200':
          description: Submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransitionResult' }
        '422': { $ref: '#/components/responses/TransitionRejected' }
...
  /applications/{applicationId}/transitions:
    post:
      tags: [Workflow]
      summary: Move an application to another stage
      description: |
        The server looks up the configured transition, checks the caller's role,
        evaluates every guard, then runs actions in one transaction.
        If any guard fails, nothing changes and `422` lists each failure.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransitionRequest' }
      responses:
        '200':
          description: Transition applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransitionResult' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/TransitionRejected' }
  /applications/{applicationId}/transitions/available:
    get:
      tags: [Workflow]
      summary: Which stages the caller can move to, with live guard status
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
      responses:
        '200':
          description: Options
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    toStage: { type: string, example: REGISTRATION }
                    allowed: { type: boolean }
                    guards:
                      type: array
                      items:
                        type: object
                        properties:
                          key: { type: string, example: DOCS_VERIFIED }
                          passed: { type: boolean }
                          message: { type: string, nullable: true }
...
  /applications/{applicationId}/documents:
    parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
    get:
      tags: [Documents]
      summary: List documents and required-document checklist
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
    post:
      tags: [Documents]
      summary: Request a pre-signed S3 upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [docType, fileName, contentType, sizeBytes]
              properties:
                docType: { type: string, example: BIRTH_CERT }
                fileName: { type: string }
                contentType: { type: string, enum: [application/pdf, image/jpeg, image/png] }
                sizeBytes: { type: integer, maximum: 5242880 }
      responses:
        '201':
          description: Upload target
          content:
            application/json:
              schema:
                type: object
                properties:
                  documentId: { type: string, format: uuid }
                  uploadUrl: { type: string, format: uri }
                  expiresInSeconds: { type: integer, example: 300 }
  /documents/{documentId}/verification:
    put:
      tags: [Documents]
      summary: Verify or reject a document (staff)
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [VERIFIED, REJECTED] }
                reason: { type: string, description: Required when REJECTED }
      responses:
        '200': { description: Updated }

  # -------------------------------------------------------------- Scheduling
...
  /applications/{applicationId}/assessments:
    post:
      tags: [Assessment]
      summary: Record interaction or test scores against the class rubric
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, scores]
              properties:
                kind: { type: string, enum: [INTERACTION, ENTRANCE_TEST] }
                remarks: { type: string }
                scores:
                  type: array
                  items:
                    type: object
                    required: [rubricId, score]
                    properties:
                      rubricId: { type: string, format: uuid }
                      score: { type: number }
      responses:
        '201':
          description: Weighted total and pass/fail computed server-side
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalScore: { type: number, example: 72.5 }
                  passed: { type: boolean }

  # ---------------------------------------------------------------- Payments
...
  /forms:
    get:
      tags: [Forms and config]
      summary: List form definitions
      responses:
        '200':
          description: Forms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FormDefinition' }
    post:
      tags: [Forms and config]
      summary: Create a form draft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FormDefinition' }
      responses:
        '201': { description: Created }
  /forms/{formId}/publish:
    post:
      tags: [Forms and config]
      summary: Publish a form (new immutable version; existing applications keep their version)
      parameters:
        - { name: formId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Published }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The application changed since you loaded it (stale `If-Match` version)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Field validation failed
      content:
        application/problem+json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Problem'
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field: { type: string }
                        message: { type: string }
    TransitionRejected:
      description: One or more guards failed; nothing was changed
      content:
        application/problem+json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Problem'
              - type: object
                properties:
                  failures:
                    type: array
                    items: { type: string }
                    example:
                      - 2 documents not verified yet (Address proof, Previous report card).
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
    Page:
      type: object
      properties:
        items: { type: array, items: {} }
        nextCursor: { type: string, nullable: true }
...
    Applicant:
      type: object
      required: [fullName, dob]
      properties:
        fullName: { type: string }
        dob: { type: string, format: date }
        gender: { type: string }
        previousSchool: { type: string }
    Guardian:
      type: object
      required: [relation, fullName, phone]
      properties:
        relation: { type: string, example: Mother }
        fullName: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        isPrimary: { type: boolean }
    Application:
      type: object
      properties:
        id: { type: string, format: uuid }
        applicationNo: { type: string, example: APP27-000123 }
        classCode: { type: string }
        stage: { type: string, example: APPLICATION }
        version: { type: integer, format: int64 }
        admissionNo: { type: string, nullable: true }
        applicant: { $ref: '#/components/schemas/Applicant' }
        assignedTo: { type: string, format: uuid, nullable: true }
        submittedAt: { type: string, format: date-time, nullable: true }
    ApplicationDetail:
      allOf:
        - $ref: '#/components/schemas/Application'
        - type: object
          properties:
            guardians:
              type: array
              items: { $ref: '#/components/schemas/Guardian' }
            responses:
              type: object
              additionalProperties: true
            documents:
              type: array
              items: { $ref: '#/components/schemas/Document' }
            nextActions:
              type: array
              items: { type: string }
              description: Plain-language steps for the guardian, driven by the current stage
    TransitionRequest:
      type: object
      required: [toStage]
      properties:
        toStage: { type: string, example: REGISTRATION }
        reason: { type: string, description: Required for WAITLISTED and REJECTED }
        expectedVersion: { type: integer, format: int64 }
    TransitionResult:
      type: object
      properties:
        applicationId: { type: string, format: uuid }
        fromStage: { type: string }
        toStage: { type: string }
        version: { type: integer, format: int64 }
        actionsRun:
          type: array
          items: { type: string }
          example: [HOLD_SEAT, ISSUE_INVOICE, NOTIFY]
...
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        docType: { type: string }
        fileName: { type: string }
        verification: { type: string, enum: [MISSING, UPLOADED, VERIFIED, REJECTED] }
        rejectReason: { type: string, nullable: true }
...
    FormDefinition:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        purpose: { type: string, enum: [ENQUIRY, APPLICATION] }
        version: { type: integer }
        status: { type: string, enum: [DRAFT, PUBLISHED, ARCHIVED] }
        fields:
          type: array
          items:
            type: object
            required: [fieldKey, label, fieldType, sequence]
            properties:
              fieldKey: { type: string }
              label: { type: string }
              fieldType: { type: string, enum: [TEXT, NUMBER, DATE, SELECT, MULTISELECT, PHONE, EMAIL, FILE, CHECKBOX, TEXTAREA] }
              isMandatory: { type: boolean }
              sequence: { type: integer }
              options:
                type: array
                items: { type: string }
...